// src/types/connection.ts
export interface Connection {
    originId:         number;
    originLatitude:   string;
    originLongitude:  string;
    weight:           number;
    destinyId:        number;
    destinyLatitude:  string;
    destinyLongitude: string;
    originName:       string;
    destinyName:      string;
}